Enum slint_build::EmbedResourcesKind

pub enum EmbedResourcesKind {
    AsAbsolutePath,
    EmbedFiles,
    EmbedForSoftwareRenderer,
}

How should the slint compiler embed images and fonts

Parameter of CompilerConfiguration::embed_resources()

Variants

AsAbsolutePath

Paths specified in .slint files are made absolute and the absolute paths will be used at run-time to load the resources from the file system.

EmbedFiles

The raw files in .slint files are embedded in the application binary.

EmbedForSoftwareRenderer

File names specified in .slint files will be loaded by the Slint compiler, optimized for use with the software renderer and embedded in the application binary.