slint_build
Struct CompilerConfiguration

pub struct CompilerConfiguration { /* private fields */ }

The structure for configuring aspects of the compilation of .slint markup files to Rust.

Implementations

impl CompilerConfiguration

pub fn new() -> Self

Creates a new default configuration.

pub fn with_include_paths(self, include_paths: Vec<PathBuf>) -> Self

Create a new configuration that includes sets the include paths used for looking up .slint imports to the specified vector of paths.

pub fn with_library_paths(self, library_paths: HashMap<String, PathBuf>) -> Self

Create a new configuration that sets the library paths used for looking up @library imports to the specified map of paths.

Each library path can either be a path to a .slint file or a directory. If it’s a file, the library is imported by its name prefixed by @ (e.g. @example). The specified file is the only entry-point for the library and other files from the library won’t be accessible from the outside. If it’s a directory, a specific file in that directory must be specified when importing the library (e.g. @example/widgets.slint). This allows exposing multiple entry-points for a single library.

Compile ui/main.slint and specify an “example” library path:

let manifest_dir = std::path::PathBuf::from(std::env::var_os("CARGO_MANIFEST_DIR").unwrap());
let library_paths = std::collections::HashMap::from([(
    "example".to_string(),
    manifest_dir.join("third_party/example/ui/lib.slint"),
)]);
let config = slint_build::CompilerConfiguration::new().with_library_paths(library_paths);
slint_build::compile_with_config("ui/main.slint", config).unwrap();

Import the “example” library in ui/main.slint:

import { Example } from "@example";

pub fn with_style(self, style: String) -> Self

Create a new configuration that selects the style to be used for widgets.

pub fn embed_resources(self, kind: EmbedResourcesKind) -> Self

Selects how the resources such as images and font are processed.

See EmbedResourcesKind

pub fn with_scale_factor(self, factor: f32) -> Self

Sets the scale factor to be applied to all px to phx conversions as constant value. This is only intended for MCU environments. Use in combination with Self::embed_resources to pre-scale images and glyphs accordingly.