slint_build

Struct CompilerConfiguration

pub struct CompilerConfiguration { /* private fields */ }

The structure for configuring aspects of the compilation of .slint markup files to Rust.

Implementations

impl CompilerConfiguration

pub fn new() -> Self

Creates a new default configuration.

pub fn with_include_paths(self, include_paths: Vec<PathBuf>) -> Self

Create a new configuration that includes sets the include paths used for looking up .slint imports to the specified vector of paths.

pub fn with_library_paths(self, library_paths: HashMap<String, PathBuf>) -> Self

Create a new configuration that sets the library paths used for looking up @library imports to the specified map of paths.

Each library path can either be a path to a .slint file or a directory. If itโs a file, the library is imported by its name prefixed by @ (e.g. @example). The specified file is the only entry-point for the library and other files from the library wonโt be accessible from the outside. If itโs a directory, a specific file in that directory must be specified when importing the library (e.g. @example/widgets.slint). This allows exposing multiple entry-points for a single library.

Compile ui/main.slint and specify an โexampleโ library path:

let manifest_dir = std::path::PathBuf::from(std::env::var_os("CARGO_MANIFEST_DIR").unwrap());
let library_paths = std::collections::HashMap::from([(
    "example".to_string(),
    manifest_dir.join("third_party/example/ui/lib.slint"),
)]);
let config = slint_build::CompilerConfiguration::new().with_library_paths(library_paths);
slint_build::compile_with_config("ui/main.slint", config).unwrap();

Import the โexampleโ library in ui/main.slint:

import { Example } from "@example";

pub fn with_style(self, style: String) -> Self

Create a new configuration that selects the style to be used for widgets.

pub fn embed_resources(self, kind: EmbedResourcesKind) -> Self

Selects how the resources such as images and font are processed.

See EmbedResourcesKind

pub fn with_scale_factor(self, factor: f32) -> Self

Sets the scale factor to be applied to all px to phx conversions as constant value. This is only intended for MCU environments. Use in combination with Self::embed_resources to pre-scale images and glyphs accordingly.

If this is set, changing the scale factor at runtime will not have any effect.

pub fn with_bundled_translations( self, path: impl Into<PathBuf>, ) -> CompilerConfiguration

Configures the compiler to bundle translations when compiling Slint code.

It expects the path to be the root directory of the translation files.

If given a relative path, it will be resolved relative to $CARGO_MANIFEST_DIR.

The translation files should be in the gettext .po format and follow this pattern: <path>/<lang>/LC_MESSAGES/<crate>.po

pub fn with_default_translation_context( self, default_translation_context: DefaultTranslationContext, ) -> Self

Unless explicitly specified with the @tr("context" => ...), the default translation context is the component name. Use this option with DefaultTranslationContext::None to disable the default translation context.

The translation file must also not have context (--no-default-translation-context argument of slint-tr-extractor)

pub fn as_library(self, library_name: &str) -> Self

Configures the compiler to treat the Slint as part of a library.

Use this when the components and types of the Slint code need to be accessible from other modules.

Note: This feature is experimental and may change or be removed in the future.

pub fn rust_module(self, rust_module: &str) -> Self

Specify the Rust module to place the generated code in.

Note: This feature is experimental and may change or be removed in the future.

pub fn with_sdf_fonts(self, enable: bool) -> Self

Configures the compiler to use Signed Distance Field (SDF) encoding for fonts.

This flag only takes effect when embed_resources is set to EmbedResourcesKind::EmbedForSoftwareRenderer, and requires the sdf-fonts cargo feature to be enabled.

SDF reduces the binary size by using an alternative representation for fonts, trading off some rendering quality for a smaller binary footprint. Rendering is slower and may result in slightly inferior visual output. Use this on systems with limited flash memory.